Use issuing company VAT rate in order offer calculation
Resolve VAT percent from BillingCompany flagged as invoice issuer via
IssuingCompanyResolver, fallback to TAX_VAT. Persist VAT amount and fee on
offers from admin/command; DTO vatPercent is float; fix OrderOffer __toString.

Made-with: Cursor

+++ src/Service/OrderOffer/DTO/OrderOfferCalculationResultDto.php
@@ -32,7 +32,7 @@ final readonly class OrderOfferCalculationResultDto
         public ?string $currency = null,
         public ?int    $bruttoPrice = null,
         public ?int    $nettoPrice = null,
-        public ?int    $vatPercent = null,
+        public ?float  $vatPercent = null,
         public ?int    $vatAmount = null,
         public ?int    $feeAmount = null,
         public ?string $errorMessage = null,
@@ -48,7 +48,7 @@ final readonly class OrderOfferCalculationResultDto
         string $currency,
         int    $bruttoPrice,
         int    $nettoPrice,
-        int    $vatPercent,
+        float  $vatPercent,
         int    $vatAmount,
         int    $feeAmount,
     ): self {
